CA-1654 Moving Slick classes around for BillingAccountChanges so they exist with the RawlsBillingProjectComponent
