+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/slick/DataAccess.scala
@@ -1,7 +1,6 @@
 package org.broadinstitute.dsde.rawls.dataaccess.slick
 
 import org.broadinstitute.dsde.rawls.entities.local.LocalEntityExpressionQueries
-import org.broadinstitute.dsde.rawls.monitor.BillingAccountChanges
 import org.broadinstitute.dsde.rawls.monitor.migration.WorkspaceMigrationActor
 import slick.jdbc.JdbcProfile
 
@@ -74,7 +73,7 @@ trait DataAccess
       TableQuery[CloneWorkspaceFileTransferTable].delete andThen  // FK to workspace
       WorkspaceMigrationActor.truncate andThen                    // FK to workspace
       TableQuery[WorkspaceTable].delete andThen
-      BillingAccountChanges.truncate andThen                      // FK to BillingProject
+      TableQuery[BillingAccountChanges].delete andThen            // FK to BillingProject
       TableQuery[RawlsBillingProjectTable].delete andThen
       TableQuery[WorkflowAuditStatusTable].delete andThen
       TableQuery[SubmissionAuditStatusTable].delete andThen
