+++ core/src/main/scala/org/broadinstitute/dsde/rawls/dataaccess/slick/RawlsBillingProjectComponent.scala
@@ -1,12 +1,20 @@
 package org.broadinstitute.dsde.rawls.dataaccess.slick
 
+import cats.implicits.catsSyntaxOptionId
 import org.broadinstitute.dsde.rawls.RawlsException
 import org.broadinstitute.dsde.rawls.dataaccess.GoogleApiTypes.GoogleApiType
 import org.broadinstitute.dsde.rawls.dataaccess.GoogleOperationNames.GoogleOperationName
 import org.broadinstitute.dsde.rawls.dataaccess.{GoogleApiTypes, GoogleOperationNames}
 import org.broadinstitute.dsde.rawls.model.CreationStatuses.CreationStatus
 import org.broadinstitute.dsde.rawls.model._
+import org.broadinstitute.dsde.rawls.monitor.migration.MigrationUtils
+import org.broadinstitute.dsde.rawls.monitor.migration.MigrationUtils.Implicits.{IgnoreResultExtensionMethod, InsertExtensionMethod}
+import org.broadinstitute.dsde.rawls.monitor.migration.MigrationUtils.Outcome
 import org.broadinstitute.dsde.workbench.model.google.{BigQueryDatasetName, BigQueryTableName, GoogleProject}
+
+import java.sql.Timestamp
+import java.time.Instant
+import scala.concurrent.ExecutionContext
 
 case class RawlsBillingProjectRecord(projectName: String, creationStatus: String, billingAccount: Option[String], message: Option[String], cromwellBackend: Option[String], servicePerimeter: Option[String], googleProjectNumber: Option[String], invalidBillingAccount: Boolean, spendReportDataset: Option[String], spendReportTable: Option[String], spendReportDatasetGoogleProject: Option[String])
 case class RawlsBillingProjectOperationRecord(projectName: String, operationName: GoogleOperationName, operationId: String, done: Boolean, errorMessage: Option[String], api: GoogleApiType)
@@ -30,6 +38,29 @@ trait RawlsBillingProjectComponent {
     def spendReportDatasetGoogleProject = column[Option[String]]("SPEND_REPORT_DATASET_GOOGLE_PROJECT", O.Length(1024))
 
     def * = (projectName, creationStatus, billingAccount, message, cromwellBackend, servicePerimeter, googleProjectNumber, invalidBillingAccount, spendReportDataset, spendReportTable, spendReportDatasetGoogleProject) <> (RawlsBillingProjectRecord.tupled, RawlsBillingProjectRecord.unapply)
+  }
+
+  final class BillingAccountChanges(tag: Tag)
+    extends Table[BillingAccountChange](tag, "BILLING_ACCOUNT_CHANGES") {
+
+    def id = column[Long]("ID", O.PrimaryKey, O.AutoInc)
+    def billingProjectName = column[String]("BILLING_PROJECT_NAME")
+    def userId = column[String]("USER_ID")
+    def originalBillingAccount = column[Option[String]]("ORIGINAL_BILLING_ACCOUNT")
+    def newBillingAccount = column[Option[String]]("NEW_BILLING_ACCOUNT")
+    def created = column[Timestamp]("CREATED")
+    def googleSyncTime = column[Option[Timestamp]]("GOOGLE_SYNC_TIME")
+    def outcome = column[Option[String]]("OUTCOME")
+    def message = column[Option[String]]("MESSAGE")
+
+    override def * =
+      (
+        id, billingProjectName, userId, originalBillingAccount, newBillingAccount, created,
+        googleSyncTime, outcome, message
+      ) <> (
+        r => MigrationUtils.unsafeFromEither(BillingAccountChanges.fromRecord(r)),
+        BillingAccountChanges.toRecord(_: BillingAccountChange).some
+      )
   }
 
   // these 2 implicits are lazy because there is a timing problem initializing MappedColumnType, if they are not lazy
@@ -191,5 +222,81 @@ trait RawlsBillingProjectComponent {
       BillingProjectSpendExport(RawlsBillingProjectName(projectRecord.projectName), billingAccount, table)
     }
   }
+
+  object BillingAccountChanges {
+
+    type RecordType = (
+      Long, // id
+        String, // billing project name
+        String, // User Id
+        Option[String], // Original billing account
+        Option[String], // New billing account
+        Timestamp, // Created
+        Option[Timestamp], // Google sync time
+        Option[String], // Outcome
+        Option[String] // Message
+      )
+
+    def fromRecord(record: RecordType): Either[String, BillingAccountChange] = record match {
+      case (id, billingProjectName, userId, originalBillingAccount, newBillingAccount, created, googleSyncTime, outcome, message) =>
+        Outcome.fromFields(outcome, message).map { outcome =>
+          BillingAccountChange(
+            id,
+            RawlsBillingProjectName(billingProjectName),
+            RawlsUserSubjectId(userId),
+            originalBillingAccount.map(RawlsBillingAccountName),
+            newBillingAccount.map(RawlsBillingAccountName),
+            created.toInstant,
+            googleSyncTime.map(_.toInstant),
+            outcome
+          )
+        }
+    }
+
+    def toRecord(billingAccountChange: BillingAccountChange): RecordType = {
+      val (outcome, message) = billingAccountChange.outcome.map(Outcome.toTuple).getOrElse((None, None))
+      (
+        billingAccountChange.id,
+        billingAccountChange.billingProjectName.value,
+        billingAccountChange.userId.value,
+        billingAccountChange.originalBillingAccount.map(_.value),
+        billingAccountChange.newBillingAccount.map(_.value),
+        Timestamp.from(billingAccountChange.created),
+        billingAccountChange.googleSyncTime.map(Timestamp.from),
+        outcome,
+        message
+      )
+    }
+  }
+
+  object billingAccountChangeQuery extends TableQuery(new BillingAccountChanges(_)) {
+    def create(billingProjectName: RawlsBillingProjectName,
+               oldBillingAccount: Option[RawlsBillingAccountName],
+               newBillingAccount: Option[RawlsBillingAccountName],
+               userId: RawlsUserSubjectId): ReadWriteAction[Unit] =
+      billingAccountChangeQuery
+        .map(change => (change.billingProjectName, change.originalBillingAccount, change.newBillingAccount, change.userId))
+        .insert((billingProjectName.value, oldBillingAccount.map(_.value), newBillingAccount.map(_.value), userId.value))
+        .ignore
+
+    def lastChange(billingProjectName: RawlsBillingProjectName)
+                  (implicit executionContext: ExecutionContext): ReadWriteAction[Option[BillingAccountChange]] =
+      billingAccountChangeQuery
+        .filter(_.billingProjectName === billingProjectName.value)
+        .sortBy(_.id.desc)
+        .take(1)
+        .result
+        .map(_.headOption)
+  }
 }
+
+final case class BillingAccountChange(id: Long,
+                                      billingProjectName: RawlsBillingProjectName,
+                                      userId: RawlsUserSubjectId,
+                                      originalBillingAccount: Option[RawlsBillingAccountName],
+                                      newBillingAccount: Option[RawlsBillingAccountName],
+                                      created: Instant,
+                                      googleSyncTime: Option[Instant],
+                                      outcome: Option[Outcome]
+                                     )
